+++ src/test/java/org/folio/inventory/dataimport/consumers/MarcBibUpdateConsumerVerticleTest.java
@@ -1,17 +1,4 @@
-package org.folio.inventory.dataimport.consumers.vertices;
-
-import io.vertx.core.DeploymentOptions;
-import io.vertx.core.Promise;
-import io.vertx.core.Vertx;
-import io.vertx.core.json.JsonObject;
-import io.vertx.ext.unit.Async;
-import io.vertx.ext.unit.TestContext;
-import io.vertx.ext.unit.junit.VertxUnitRunner;
-import net.mguenther.kafka.junit.EmbeddedKafkaCluster;
-import org.folio.inventory.MarcHridSetConsumerVerticle;
-import org.junit.AfterClass;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+package org.folio.inventory.dataimport.consumers;
 
 import static net.mguenther.kafka.junit.EmbeddedKafkaCluster.provisionWith;
 import static net.mguenther.kafka.junit.EmbeddedKafkaClusterConfig.defaultClusterConfig;
@@ -21,14 +8,36 @@ import static org.folio.inventory.dataimport.util.KafkaConfigConstants.KAFKA_MAX
 import static org.folio.inventory.dataimport.util.KafkaConfigConstants.KAFKA_PORT;
 import static org.folio.inventory.dataimport.util.KafkaConfigConstants.KAFKA_REPLICATION_FACTOR;
 
-@RunWith(VertxUnitRunner.class)
-public class MarcHridSetConsumerVerticleTest {
+import io.vertx.core.DeploymentOptions;
+import io.vertx.core.Promise;
+import io.vertx.core.Vertx;
+import io.vertx.core.json.JsonObject;
+import io.vertx.ext.unit.Async;
+import io.vertx.ext.unit.TestContext;
+import io.vertx.ext.unit.junit.VertxUnitRunner;
+import net.mguenther.kafka.junit.EmbeddedKafkaCluster;
+import org.folio.inventory.MarcBibUpdateConsumerVerticle;
+import org.folio.inventory.dataimport.consumers.MarcBibUpdateKafkaHandler;
+import org.junit.AfterClass;
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.Mock;
+import org.mockito.MockitoAnnotations;
 
-  private static final String TENANT_ID = "diku";
+@RunWith(VertxUnitRunner.class)
+public class MarcBibUpdateConsumerVerticleTest {
   private static final String KAFKA_ENV_NAME = "test-env";
   private static Vertx vertx = Vertx.vertx();
-
   public static EmbeddedKafkaCluster cluster;
+
+  @Mock
+  private static MarcBibUpdateKafkaHandler marcBibUpdateKafkaHandler;
+
+  @Before
+  public void setUp() {
+    MockitoAnnotations.openMocks(this);
+  }
 
   @Test
   public void shouldDeployVerticle(TestContext context) {
@@ -46,7 +55,7 @@ public class MarcHridSetConsumerVerticleTest {
       .setWorker(true);
 
     Promise<String> promise = Promise.promise();
-    vertx.deployVerticle(MarcHridSetConsumerVerticle.class.getName(), options, promise);
+    vertx.deployVerticle(MarcBibUpdateConsumerVerticle.class.getName(), options, promise);
 
     promise.future().onComplete(ar -> {
       context.assertTrue(ar.succeeded());
